src/processor/requirements.txt
azure-eventhub
azure-eventhub-checkpointstoreblob-aio
azure-storage-file-datalake
sqlalchemy
psycopg2-binary